Code written to  monitor Linux cron daemons during a total datacenter shutdown.
Various Linux crons like Depot.sh/Automount.sh and Ping check usually
are the setermining factors towards the wellness of a machine and the
machine is said to be in a good state of health if the cron daemons are
working absolutely fine. So the c ode wriiten is to monitor their
behaviour and whether all these daemons pass the checks as enlisted in
the code.